the old hfsm was not working as expected.
the hfsm is a non deterministic state machine. Meaning, it can land on multiple dst states from the same src state. And the only reason it lands on one out of many dst states is the guard. For this to work properly, only the guard of the intended dst state should return True, and guards of other possible dst states should return False.

This trick was taken intro consideration into this commit.